:recycle: move common model definition in shared models
