+++ src/app/samples/samples.model.ts
@@ -1,26 +1,9 @@
-export interface ObjectID {
-  $oid: string;
-}
+import { JSONObject, ObjectID } from "../shared/shared.model";
 
 export interface Multipoint {
   coordinates: number[][];
   type: string;
 }
-
-// https://dev.to/ankittanna/how-to-create-a-type-for-complex-json-object-in-typescript-d81
-export type JSONValue =
-    | string
-    | number
-    | boolean
-    | JSONObject
-    | JSONArray;
-
-export interface JSONObject {
-  [x: string]: JSONValue;
-}
-
-export interface JSONArray extends Array<JSONValue> { }
-
 export interface Sample {
   _id: ObjectID;
   alias: string;
